requirements.txt
asgiref==3.5.2
astroid==2.12.13
autopep8==2.0.0
coverage==6.5.0
dill==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.5.4
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.15.7
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
tzdata==2022.6
wrapt==1.14.1
